requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
jinja2==3.1.3
python-multipart==0.0.9
python-docx==1.1.0
pymupdf==1.23.22
pillow==10.3.0
pytesseract==0.3.10
python-dotenv==1.0.1
fpdf==1.7.2